refactor(schemas): publish forced barrel leaves
